requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
antlr4-python3-runtime @ file:///home/conda/feedstock_root/build_artifacts/antlr-python-runtime-meta_1638309185939/work
anyio==4.5.2
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==5.0.1
attrs==25.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
blis==0.7.11
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
brotlipy==0.7.0
cachetools==5.5.1
catalogue==2.0.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
confection==0.1.5
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1686613057838/work
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.12
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
diffusers==0.32.2
distro==1.9.0
einops==0.8.0
enchant==0.0.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup==1.2.2
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.2.3
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hydra-core==1.3.2
idna @ file:///croot/idna_1666125576474/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython @ file:///croot/ipython_1691532092695/work
ipython-genutils==0.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1690896916983/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.3.0
lightning-utilities==0.11.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.3.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.1.0
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1690561992265/work
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
omegaconf @ file:///home/conda/feedstock_root/build_artifacts/omegaconf_1670575376789/work
openai==0.28.0
orjson==3.10.15
packaging==24.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathlib_abc==0.1.1
pathy==0.11.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.3.0
platformdirs==4.3.6
preshed==3.0.9
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
propcache==0.2.0
psutil==6.1.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.21
pydantic_core==2.27.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
PyJWT==2.8.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-lightning==2.4.0
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.2.1
regex==2024.11.6
requests @ file:///croot/requests_1721410876868/work
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
smplx==0.1.28
sniffio==1.3.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1728484454203/work
tenacity==9.0.0
thinc==8.1.12
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchmetrics==1.5.2
torchvision==0.18.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.46.3
triton==2.3.1
typer==0.7.0
typing_extensions==4.12.2
urllib3 @ file:///croot/urllib3_1727769808118/work
wasabi==1.1.3
wcwidth==0.2.13
yarl==1.15.2
zhipuai==2.1.5.20250106
zipp==3.20.2
